packages/auth/pyproject.toml
[project]
name = "nexus-care-auth"
version = "0.1.0"
description = "Authentication primitives — Argon2id PIN hashing, JWT, permissions."
requires-python = ">=3.13,<3.14"
dependencies = [
    "argon2-cffi>=23.1.0",
    "pyjwt>=2.10.0",
    "pydantic>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus_care_auth"]
